refactor: Group recommendation tools.
