+++ src/tools/index.ts
@@ -9,7 +9,6 @@ import "./create-report-notification";
 import "./create-virtual-tag-config";
 import "./delete-cost-report";
 import "./delete-folder";
-import "./delete-recommendation-view";
 import "./delete-report-notification";
 import "./get-anomaly";
 import "./get-cost-alert";
@@ -20,9 +19,6 @@ import "./get-financial-commitment-report";
 import "./get-folder";
 import "./get-myself";
 import "./get-provider-resource";
-import "./get-recommendation-details";
-import "./get-recommendation-resource-details";
-import "./get-recommendation-resources";
 import "./get-report-notification";
 import "./get-team";
 import "./get-teams";
@@ -38,12 +34,12 @@ import "./list-costs";
 import "./list-dashboards";
 import "./list-folders";
 import "./list-provider-resources";
-import "./list-recommendations";
 import "./list-report-notifications";
 import "./list-tag-values";
 import "./list-tags";
 import "./list-unit-costs";
 import "./query-costs";
+import "./recommendations";
 import "./submit-user-feedback";
 import "./update-anomaly";
 import "./update-folder";
