+++ src/tools/recommendations/index.ts
@@ -1,0 +1,5 @@
+import "./delete-recommendation-view";
+import "./get-recommendation-details";
+import "./get-recommendation-resource-details";
+import "./get-recommendation-resources";
+import "./list-recommendations";
